BENCH: collect low-dependency asvs
