+++ asv_bench/benchmarks/algorithms.py
@@ -2,8 +2,6 @@ from importlib import import_module
 
 import numpy as np
 
-from pandas._libs import lib
-
 import pandas as pd
 
 from .pandas_vb_common import tm
@@ -14,19 +12,6 @@ for imp in ["pandas.util", "pandas.tools.hashing"]:
         break
     except (ImportError, TypeError, ValueError):
         pass
-
-
-class MaybeConvertObjects:
-    def setup(self):
-        N = 10 ** 5
-
-        data = list(range(N))
-        data[0] = pd.NaT
-        data = np.array(data)
-        self.data = data
-
-    def time_maybe_convert_objects(self):
-        lib.maybe_convert_objects(self.data)
 
 
 class Factorize:
