+++ asv_bench/benchmarks/reindex.py
@@ -9,10 +9,7 @@ from pandas import (
     period_range,
 )
 
-from .pandas_vb_common import (
-    lib,
-    tm,
-)
+from .pandas_vb_common import tm
 
 
 class Reindex:
@@ -153,21 +150,6 @@ class Align:
 
     def time_align_series_irregular_string(self):
         self.x + self.y
-
-
-class LibFastZip:
-    def setup(self):
-        N = 10000
-        K = 10
-        key1 = tm.makeStringIndex(N).values.repeat(K)
-        key2 = tm.makeStringIndex(N).values.repeat(K)
-        col_array = np.vstack([key1, key2, np.random.randn(N * K)])
-        col_array2 = col_array.copy()
-        col_array2[:, :10000] = np.nan
-        self.col_array_list = list(col_array)
-
-    def time_lib_fast_zip(self):
-        lib.fast_zip(self.col_array_list)
 
 
 from .pandas_vb_common import setup  # noqa: F401 isort:skip
